+++ frontend/test/e2e/CreateNodeValidator.spec.ts
@@ -1,19 +1,19 @@
 import { WebDriver } from 'selenium-webdriver';
 
-import { SettingsPage } from '../pages/SettingsPage.js';
+import { ValidatorsPage } from '../pages/ValidatorsPage.js';
 import { ContractsPage } from '../pages/ContractsPage.js';
 import { before, describe, after, it } from 'node:test';
 import { expect } from 'chai';
 import { getDriver } from '../utils/driver.js';
 
 let driver: WebDriver;
-let settingsPage: SettingsPage;
+let validatorsPage: ValidatorsPage;
 let contractsPage: ContractsPage;
 
 describe('Settings - Create Node Validator', () => {
   before(async () => {
     driver = await getDriver();
-    settingsPage = new SettingsPage(driver);
+    validatorsPage = new ValidatorsPage(driver);
     contractsPage = new ContractsPage(driver);
   });
 
@@ -22,16 +22,16 @@ describe('Settings - Create Node Validator', () => {
     await contractsPage.waitUntilVisible();
     await contractsPage.skipTutorial();
 
-    await settingsPage.navigate();
-    await settingsPage.waitUntilVisible();
+    await validatorsPage.navigate();
+    await validatorsPage.waitUntilVisible();
 
-    const initialValidators = await settingsPage.getValidatorsElements();
-    await settingsPage.createValidator({
+    const initialValidators = await validatorsPage.getValidatorsElements();
+    await validatorsPage.createValidator({
       provider: 'heuristai',
       model: 'mistralai/mixtral-8x7b-instruct',
       stake: 7,
     });
-    const existingValidators = await settingsPage.getValidatorsElements();
+    const existingValidators = await validatorsPage.getValidatorsElements();
     expect(
       existingValidators.length,
       'number of validators should be greather than old validators list',
